requirements.txt
requests>=2.31.0
SpeechRecognition>=3.10.0
pyautogui>=0.9.54
psutil>=5.9.0
python-dotenv>=1.0.0
openai>=1.0.0
keyboard>=0.13.5
pillow>=10.0.0
edge-tts>=7.0.0
playsound==1.2.2
sounddevice>=0.5.0
numpy>=2.0.0
pyinstaller>=6.0.0
pystray>=0.19.0
pyperclip>=1.8.0
pygame>=2.5.0
google-generativeai>=0.7.0
